Authentication and authorization, including tests

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -12,6 +13,7 @@ class CommentController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Comment::class);
         return Comment::all();
     }
 
@@ -28,6 +30,7 @@ class CommentController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Comment::class);
         $comment = Comment::create($request->all());
         return response()->json($comment, 201);
     }
@@ -37,7 +40,9 @@ class CommentController extends Controller
      */
     public function show(string $id)
     {
-        return Comment::findorFail($id);
+        $comment = Comment::findorFail($id);
+        Gate::authorize('view', $comment);
+        return $comment;
     }
 
     /**
@@ -54,6 +59,7 @@ class CommentController extends Controller
     public function update(Request $request, string $id)
     {
         $comment = Comment::findorFail($id);
+        Gate::authorize('update', $comment);
         $comment->update($request->all());
         return response()->json($comment, 200);
     }
@@ -63,6 +69,8 @@ class CommentController extends Controller
      */
     public function destroy(string $id)
     {
+        $comment = Comment::findorFail($id);
+        Gate::authorize('delete', $comment);
         Comment::destroy($id);
         return response()->json(null, 204);
     }
